Fixed when kit has use limit, but no time limit

+++ KitZ/Commands.cs
@@ -58,7 +58,7 @@ namespace KitZ
                             TShock.Utils.GetItemById(kitItem.Id).name));
                 }
                 var kitUse = await KitZ.Kits.GetKitUseAsync(e.Player, kit);
-                if (kitUse == null) return;
+                if (kitUse == null || kit.RefreshTime == TimeSpan.Zero) return;
                 await Task.Delay(kitUse.ExpireTime - DateTime.UtcNow);
                 await KitZ.Kits.DeleteKitUseAsync(kitUse);
             }
